refactoring

+++ src/usecases/user_tracking/remove_user_from_tracking.py
@@ -1,12 +1,29 @@
+from dto import RemoveUserTrackingDTO
 from repositories import UserTrackingRepository
+from services import UserService
 
 
 class RemoveUserFromTrackingUseCase:
-    def __init__(self, user_tracking_repository: UserTrackingRepository):
+    def __init__(
+        self,
+        user_tracking_repository: UserTrackingRepository,
+        user_service: UserService,
+    ):
         self.user_tracking_repository = user_tracking_repository
+        self.user_service = user_service
 
-    async def execute(self, admin_username: str, user_username: str) -> bool:
+    async def execute(self, dto: RemoveUserTrackingDTO) -> bool:
         """Удаляет пользователя из списка отслеживания админа."""
-        return await self.user_tracking_repository.remove_user_from_tracking(
-            admin_username=admin_username, user_username=user_username
+        admin = await self.user_service.get_user(
+            tg_id=dto.admin_tgid,
+            username=dto.admin_username,
         )
+
+        if not admin:
+            return False
+
+        await self.user_tracking_repository.remove_user_from_tracking(
+            admin_id=admin.id,
+            user_id=dto.user_id,
+        )
+        return True
